// ClientApp/apps/vendor-management/src/Modules/Shared/Enums/Approvals.enum.ts
export enum APPROVALS_COMPARISON_FILTERS {
    VENDOR = 'Vendor',
    AirportCode = 'AirportCode',
    ENTITY_NAME = 'Entity Name',
    APPROVAL_STATUS = 'Status',
    TOPUSAGEAIRPORT = 'Top Usage Airport',
    VENDORLEVELNAME = 'Vendor Level Name',
    AIRPORTRANK = 'Airport Rank',
    VENDORLOCATION = 'Vendor Location'
  }

export enum APPROVAL_MERGE_STATUS {
    NOT_MERGED = 1,
    MERGED=2,
    FAILED=3,
    COMPLETED=4,
    APPROVED_STATUS_ID = 2,
    REJECTED_STATUS_ID = 3,
  }